requirements.txt
streamlit==1.30.0
tensorflow-cpu==2.10.0
pillow
numpy

